refactor(selector): Rename selector from fullyQualifiedRouteStops to feedScopedModificationStops
