+++ lib/components/modification/__tests__/convert-to-frequency.js
@@ -1,5 +1,4 @@
 // @flow
-
 import renderer from 'react-test-renderer'
 import React from 'react'
 
@@ -12,7 +11,7 @@ describe('Component > Modification > ConvertToFrequency', () => {
     const props = {
       allPhaseFromTimetableStops: {},
       feeds: [mockFeed],
-      fullyQualifiedRouteStops: [mockStop1],
+      feedScopedModificationStops: [mockStop1],
       modification: mockModification,
       projectTimetables: [],
       routePatterns: [mockPattern],
