+++ lib/components/modification/convert-to-frequency.js
@@ -19,7 +19,7 @@ import type {
 type Props = {
   allPhaseFromTimetableStops: any,
   feeds: Feed[],
-  fullyQualifiedRouteStops: Stop[],
+  feedScopedModificationStops: Stop[],
   modification: Modification,
   routePatterns: RoutePatterns,
   projectTimetables: Timetable[],
@@ -36,8 +36,9 @@ type Props = {
  *
  * @author mattwigway
  */
-export default class ConvertToFrequencyComponent
-  extends Component<void, Props, void> {
+export default class ConvertToFrequencyComponent extends Component {
+  props: Props
+
   onRouteChange = ({
     feed,
     routes
@@ -93,7 +94,7 @@ export default class ConvertToFrequencyComponent
     const {
       allPhaseFromTimetableStops,
       feeds,
-      fullyQualifiedRouteStops,
+      feedScopedModificationStops,
       modification,
       routePatterns,
       projectTimetables,
@@ -133,7 +134,7 @@ export default class ConvertToFrequencyComponent
                 feed={selectedFeed}
                 index={eidx + 1}
                 key={eidx}
-                modificationStops={fullyQualifiedRouteStops}
+                modificationStops={feedScopedModificationStops}
                 remove={this.removeEntry.bind(this, eidx)}
                 routePatterns={routePatterns}
                 routes={modification.routes}
