+++ lib/containers/convert-to-frequency.js
@@ -9,7 +9,7 @@ import * as select from '../selectors'
 export default connect(createStructuredSelector({
   allPhaseFromTimetableStops: select.allPhaseFromTimetableStops,
   feeds: select.feedsWithBundleNames,
-  fullyQualifiedRouteStops: select.fullyQualifiedRouteStops,
+  feedScopedModificationStops: select.feedScopedModificationStops,
   projectTimetables: select.projectTimetables,
   routePatterns: select.routePatterns,
   selectedFeed: select.modificationFeed
