+++ lib/selectors/feed-scoped-modification-stops.js
