+++ lib/selectors/index.js
@@ -26,11 +26,11 @@ export {default as currentRegionId} from './current-region-id'
 export {default as currentRegion} from './current-region'
 export {default as destinationTravelTimeDistribution} from './destination-travel-time-distribution'
 export {default as extendFromEnd} from './extend-from-end'
+export {default as feedScopedModificationStops} from './feed-scoped-modification-stops'
 export {default as feedScopedStops} from './feed-scoped-stops'
 export {default as feedStopsById} from './feed-stops-by-id'
 export {default as feedStops} from './feed-stops'
 export {default as feedsWithBundleNames} from './feeds-with-bundle-names'
-export {default as fullyQualifiedRouteStops} from './fully-qualified-route-stops'
 export {default as isochrone} from './isochrone'
 export {default as lastStopDistanceFromStart} from './last-stop-distance-from-start'
 export {default as maxAccessibility} from './max-accessibility'
